Refactor for reentrancy, flexibility, and data
No more global variables; now things are passed in and handled in
a primary struct, which is reentrant.

Before, the sscanf was a bit hacky and primitive. It could not handle
when variables were empty if it expected them to be full. Now, regular
expressions can capture empty or present values and convert them
appropriately.

Added GPVTG data, which provides bearing and speed information.